+++ Entities/FinalModels/StockProfit_Final.cs
@@ -6,10 +6,11 @@ using System.Threading.Tasks;
 
 namespace Entities.FinalModels
 {
-    public class StockProfit_Final
+    public class StockProfit_Final 
     {
         public int Year { get; set; }
-        public long Profit { get; set; }
-        public float GrowthRates { get; set; } // stock profit - last year
+        public double Profit { get; set; }
+        public double GrowthRates { get; set; }
+
     }
 }
